python/grpc-server/requirements.txt
grpcio 
grpcio-tools
click>=6.6
networkx>=1.11
nltk>=3.2.1
numpy>=1.11.2
editdistance
scipy
